Major refactoring, licenses added, readme.md modified.
Split the RarExtractionInstaller in two parts, trying to mimic the
structure of the ZipExtractionInstaller and FilePath.
RarExtractionInstaller will represent the installer while the extraction
is done in RarFilePath, a modified and extended version of FilePath.
Reason: No extraction was being done on connected slaves due to parts of
the code not being serializable.

Current status:
Successful extraction on master and/or slaves from both http- and
file-addresses.
Successful extraction on both Windows 7 and Ubuntu.
Slaves are not downloading the rar file from a http address, instead
master downloads the tool, transfers tool to the slaves and start
extraction. This behaviour is the same as the ZipExtractionInstaller.

Todo:
Tests for the installer.
Modify licenses for the two classes and make sure they are correct.
Check that readme.md is correct.