Code cleanup, added unit tests
Summary:
- Code cleanup
- Added a test for RarFilePath. Extracts a sample tool archive
separately from Jenkins.
- Added a test which sets up a configured Jenkins instance with a custom
tool and a freestyle project that which should perform a automated
simulation of a tool installation with the RarExtractionInstaller. TODO:
Currently, no extraction is done. It builds all the way until it should
perform the installation in the installer, but nothing happens.
- Limitation: The installer is only capable of extracting a single RAR
file. A tool archive split into multiple volumes can not be extracted
due to the JUnrar library.